Restore UB fallback
"offsetof" requires the "stddef.h" or "cstddef" header to be included,
but that should be behind a header include check. When
"BUILTIN_TYPES_ONLY" is specified, the compiler has no extension to
query alignment and "stddef.h" or "cstddef" is not included, then this
will be used.

The cast to integral type is gone now, because you can get a "ptrdiff_t"
value by just subtracting the pointer from another member's pointer.